examples/sitegen/requirements.txt
click>=8.0.0
markdown>=3.3.0
pyyaml>=6.0
jinja2>=3.0.0
